TestHelper: fix internal SpectaTests assertions.
The current assertion macros refer to `self` (since they use XCTest
assertion macros that do that internally). This raises exceptions upon
failure reporting, since the `self` instance that is captured by these
macros is the first instance of the spec that is initialized, which is
used just to initialize the suite. The actual spec that needs to be used
is the current instance of `SPTSpec` that was created by XCTest for this
specific test case execution, which is accessible via the
`SPTCurrentSpec` macro.