SpectaDSL: avoid removal of shared data dictionary.
In case the same spec is run more than once, this will break the data
dictionary. This is negligible since in the best case, it will allow to
deallocate one dictionary per shared example, but in the common case the
shared example (which is not deallocated anytime) may hold a reference
to that dictionary anyway.